Moved schema permission tests into separate test module.
