+++ pallets/msa/src/tests/mod.rs
@@ -7,5 +7,6 @@ mod other_tests;
 mod public_key_tests;
 mod replay_tests;
 mod retirement_tests;
+mod schema_permission_tests;
 mod signature_registry_tests;
 mod signed_extension_tests;
